Extended v2

+++ app/src/pages/manage-floors-edit-page-tab-access.tsx
@@ -1,32 +1,26 @@
-import React, {useEffect, useReducer, useState} from "react";
+import React, {ChangeEvent, useEffect, useReducer, useState} from "react";
 import {Floor} from "../models/floor";
 import {
     Alert,
     AlertTitle,
-    Box,
     Button,
-    Dialog, DialogActions,
+    Dialog,
+    DialogActions,
     DialogContent,
     DialogTitle,
     List,
     ListItem,
     ListItemButton,
-    ListItemText, Paper,
-    Table,
-    TableBody,
+    ListItemText,
     TableCell,
-    TableContainer,
-    TableHead,
-    TableRow,
     TextField
 } from "@mui/material";
 import {User} from "../models/user";
 import {AccessesApiService, UserApiService} from "../services/rest-api-service";
-import {Access} from "../models/access";
-import {FormDialog} from "../dialogs/form-dialog";
 import {ConfirmDialog} from "../dialogs/confirm-dialog";
-import {AccessTable} from "../components/access-table";
 import {UserAccess} from "../models/user-access";
+import {TableView} from "../components/table-view";
+import {withDelay} from "../utils/with-delay";
 
 interface ManageFloorsEditPageTabAccessProps {
     floor: Floor;
@@ -34,41 +28,49 @@ interface ManageFloorsEditPageTabAccessProps {
 
 export function ManageFloorsEditPageTabAccess(props: ManageFloorsEditPageTabAccessProps) {
     const [showAddUserDialog, toggleShowAddUserDialog] = useReducer(p => !p, false);
-    const [accesses, setAccesses] = useState<UserAccess[]>([]);
+    const [accesses, setAccesses] = useState<UserAccess[]>();
     const [accessToRevoke, setAccessToRevoke] = useState<UserAccess>();
 
     useEffect(() => {
-        AccessesApiService.list({floor: props.floor.id.toString()})
-            .then(accessRes => {
-                Promise.all(accessRes.results.map(acc => UserApiService.retrieve(acc.user)))
-                    .then(userRes => setAccesses(accessRes.results.map((acc, index) => ({
-                        user: userRes[index],
-                        access: acc,
-                    }))));
-            })
+        withDelay(
+            () => {
+                return AccessesApiService.list({floor: props.floor.id.toString()})
+                    .then(accessRes => {
+                        return Promise.all(accessRes.results.map(acc => UserApiService.retrieve(acc.user)))
+                            .then(userRes => accessRes.results.map((acc, index) => ({
+                                user: userRes[index],
+                                access: acc,
+                            })));
+                    })
+            },
+            accesses => setAccesses(accesses),
+            500,
+        );
     }, []);
 
     const handleAddAccess = (user: User) => {
-        AccessesApiService.create({
-            floor: props.floor.id,
-            user: user.id,
-            id: 0,
-            is_maintainer: false,
-        })
-            .then(acc => {
-                setAccesses([
-                    ...accesses,
-                    {
-                        access: acc,
-                        user: user,
-                    }
-                ]);
+        if (accesses != null) {
+            AccessesApiService.create({
+                floor: props.floor.id,
+                user: user.id,
+                id: 0,
+                is_maintainer: false,
             })
-        toggleShowAddUserDialog();
+                .then(acc => {
+                    setAccesses([
+                        ...accesses,
+                        {
+                            access: acc,
+                            user: user,
+                        }
+                    ]);
+                })
+            toggleShowAddUserDialog();
+        }
     };
 
     const handleRevokeAccess = () => {
-        if (accessToRevoke != null) {
+        if (accesses != null && accessToRevoke != null) {
             AccessesApiService.destroy(accessToRevoke.access.id);
             setAccesses(accesses.filter(acc => acc.access.id !== accessToRevoke.access.id));
             setAccessToRevoke(undefined);
@@ -77,33 +79,28 @@ export function ManageFloorsEditPageTabAccess(props: ManageFloorsEditPageTabAcce
 
     return (
         <>
-            <Box
-                sx={{
-                    display: 'flex',
-                    justifyContent: 'flex-end',
-                    mb: 2,
-                }}
-            >
-                <Button
-                    variant="contained"
-                    onClick={toggleShowAddUserDialog}
-                >
-                    Nutzer Hinzufügen
-                </Button>
-            </Box>
+            <TableView
+                actions={[{
+                    label: 'Nutzer Hinzufügen',
+                    onClick: toggleShowAddUserDialog,
+                }]}
+                columLabels={['Benutzer']}
+                data={accesses}
+                cellBuilder={access => (
+                    <TableCell>
+                        {access.user.username}
+                    </TableCell>
+                )}
+                onDelete={setAccessToRevoke}
+                getRowId={access => access.access.id}
+                filterItem={(access, search) => access.user.username.toLowerCase().includes(search)}
+            />
 
             {
-                accesses.length > 0 &&
-                <AccessTable
-                    accesses={accesses}
-                    onDelete={setAccessToRevoke}
-                />
-            }
-
-            {
+                accesses != null &&
                 accesses.length === 0 &&
                 <Alert
-                    sx={{mt: 2}}
+                    sx={{mt: 4}}
                     severity="info"
                 >
                     <AlertTitle>Keine Zugangsberechtigten Nutzer</AlertTitle>
@@ -112,11 +109,15 @@ export function ManageFloorsEditPageTabAccess(props: ManageFloorsEditPageTabAcce
                 </Alert>
             }
 
-            <SelectUserDialog
-                open={showAddUserDialog}
-                onSelect={handleAddAccess}
-                onCancel={toggleShowAddUserDialog}
-            />
+            {
+                accesses != null &&
+                showAddUserDialog &&
+                <SelectUserDialog
+                    onSelect={handleAddAccess}
+                    onCancel={toggleShowAddUserDialog}
+                    existingAccesses={accesses}
+                />
+            }
 
             <ConfirmDialog
                 title="Zugang entfernen"
@@ -131,9 +132,9 @@ export function ManageFloorsEditPageTabAccess(props: ManageFloorsEditPageTabAcce
 }
 
 interface SelectUserDialogProps {
-    open: boolean;
     onSelect: (user: User) => void;
     onCancel: () => void;
+    existingAccesses: UserAccess[];
 }
 
 function SelectUserDialog(props: SelectUserDialogProps) {
@@ -142,12 +143,12 @@ function SelectUserDialog(props: SelectUserDialogProps) {
 
     useEffect(() => {
         UserApiService.list()
-            .then(res => setUsers(res.results));
+            .then(res => setUsers(res.results.filter(usr => !props.existingAccesses.some(acc => acc.user.id === usr.id))));
     }, []);
 
     return (
         <Dialog
-            open={props.open}
+            open={true}
             onClose={props.onCancel}
             fullWidth
         >
@@ -156,29 +157,44 @@ function SelectUserDialog(props: SelectUserDialogProps) {
             </DialogTitle>
 
             <DialogContent>
-                <TextField
-                    fullWidth
-                    label="Suchen"
-                    sx={{mt: 2}}
-                    value={search}
-                    onChange={evt => setSearch(evt.target.value)}
-                />
+                {
+                    users.length > 0 &&
+                    <>
+                        <TextField
+                            fullWidth
+                            label="Suchen"
+                            sx={{mt: 2}}
+                            value={search}
+                            onChange={(evt: ChangeEvent<HTMLInputElement>) => setSearch(evt.target.value)}
+                        />
 
-                <List>
-                    {
-                        users
-                            .filter(usr => usr.username.toLowerCase().includes(search.toLowerCase()))
-                            .map(usr => (
-                                <ListItem key={usr.id}>
-                                    <ListItemButton onClick={() => props.onSelect(usr)}>
-                                        <ListItemText>
-                                            {usr.username}
-                                        </ListItemText>
-                                    </ListItemButton>
-                                </ListItem>
-                            ))
-                    }
-                </List>
+                        <List>
+                            {
+                                users
+                                    .filter(usr => usr.username.toLowerCase().includes(search.toLowerCase()))
+                                    .map(usr => (
+                                        <ListItem key={usr.id}>
+                                            <ListItemButton onClick={() => props.onSelect(usr)}>
+                                                <ListItemText>
+                                                    {usr.username}
+                                                </ListItemText>
+                                            </ListItemButton>
+                                        </ListItem>
+                                    ))
+                            }
+                        </List>
+                    </>
+                }
+
+                {
+                    users.length === 0 &&
+                    <Alert
+                        severity="info"
+                        sx={{mt: 2}}
+                    >
+                        Bereits alle Nutzer hinzugefügt.
+                    </Alert>
+                }
             </DialogContent>
 
             <DialogActions>
